# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Adaptive Learning API",
    "description": "Operational endpoints plus the Phase 1 learning MVP for AWS SOA-C03.",
    "license": {
      "name": "UNLICENSED",
      "identifier": "UNLICENSED"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/certifications/{certification_id}/practice-tests": {
      "get": {
        "tags": [
          "practice-tests"
        ],
        "summary": "Lists the practice tests (exam simulations) for a certification.",
        "operationId": "list_practice_tests",
        "parameters": [
          {
            "name": "certification_id",
            "in": "path",
            "description": "Certification identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Available practice tests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeTestListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/certifications/{certification_id}/practice-tests/{practice_test_id}": {
      "get": {
        "tags": [
          "practice-tests"
        ],
        "summary": "Returns learner-safe practice-test content, before submission.",
        "description": "Canonical answers, per-choice feedback, and scored flags are never included.",
        "operationId": "get_practice_test",
        "parameters": [
          {
            "name": "certification_id",
            "in": "path",
            "description": "Certification identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "practice_test_id",
            "in": "path",
            "description": "Practice-test identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Learner-safe practice test",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeTestResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown practice test",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/certifications/{certification_id}/practice-tests/{practice_test_id}/submit": {
      "post": {
        "tags": [
          "practice-tests"
        ],
        "summary": "Scores and reviews a complete practice-test attempt.",
        "description": "The whole attempt is submitted at once; the response reveals canonical\nanswers, explanations, and per-choice feedback. The raw practice score is\nnot an AWS scaled score.",
        "operationId": "submit_practice_test",
        "parameters": [
          {
            "name": "certification_id",
            "in": "path",
            "description": "Certification identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "practice_test_id",
            "in": "path",
            "description": "Practice-test identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PracticeTestSubmissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Scored practice test",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeTestResultResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid submission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown practice test",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/missions/issue": {
      "post": {
        "tags": [
          "missions"
        ],
        "summary": "Issues a mission for a quiz mode.",
        "description": "Authenticated callers get a user-owned mission. Anonymous callers may issue\npublic demo task-practice missions only.",
        "operationId": "issue_mission",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueMissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Issued mission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown certification or task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/missions/{mission_id}/answers": {
      "post": {
        "tags": [
          "missions"
        ],
        "summary": "Scores one attempt and returns immediate feedback.",
        "operationId": "answer_mission",
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "description": "Mission identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Scored attempt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid answer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Mission belongs to another account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown mission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Mission completed or content version mismatch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/missions/{mission_id}/complete": {
      "post": {
        "tags": [
          "missions"
        ],
        "summary": "Marks a mission completed for its owner.",
        "operationId": "complete_mission",
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "description": "Mission identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteMissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Mission completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteMissionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Mission belongs to another account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown mission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/missions/{mission_id}/review": {
      "get": {
        "tags": [
          "missions"
        ],
        "summary": "Returns a read-only review of a completed mission.",
        "description": "Canonical answers are included only after completion, so in-progress work\nnever leaks answers. Review never creates evidence or changes scores.",
        "operationId": "review_mission",
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "description": "Mission identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mission review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionReviewResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Mission belongs to another account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown mission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Mission is not complete yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnswerRequest": {
        "type": "object",
        "description": "Request to score one attempt.",
        "required": [
          "event_id",
          "question_id",
          "content_version",
          "attempt_number",
          "hint_count",
          "response_ms",
          "occurred_at",
          "answer"
        ],
        "properties": {
          "answer": {
            "$ref": "#/components/schemas/AnswerPayload",
            "description": "Answer primitives."
          },
          "attempt_number": {
            "type": "integer",
            "format": "int32",
            "description": "1-based attempt number."
          },
          "content_version": {
            "type": "string",
            "description": "Content version the client believes it is answering."
          },
          "device_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Device/install context. Not an authorization proof."
          },
          "event_id": {
            "type": "string",
            "format": "uuid",
            "description": "Stable event identifier for this attempt."
          },
          "hint_count": {
            "type": "integer",
            "format": "int32",
            "description": "Hints used before submitting."
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the attempt occurred on the device."
          },
          "question_id": {
            "type": "string",
            "description": "Question being answered."
          },
          "response_ms": {
            "type": "integer",
            "format": "int32",
            "description": "Active response time in milliseconds."
          }
        }
      },
...
      "CompleteMissionRequest": {
        "type": "object",
        "description": "Request to complete a mission.\n\nThe owner is taken from the verified token, so no identity is required here.",
        "properties": {
          "device_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Device/install context. Not an authorization proof."
          }
        }
      },
      "CompleteMissionResponse": {
        "type": "object",
        "description": "Completion result.",
        "required": [
          "id",
          "status"
        ],
        "properties": {
          "completed_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Completion time."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Mission identifier."
          },
          "status": {
            "$ref": "#/components/schemas/MissionStatus",
            "description": "New status."
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Envelope returned for every API error.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody",
            "description": "Error payload."
          }
        }
      },
...
      "IssueMissionRequest": {
        "type": "object",
        "description": "Request to issue a mission for a quiz mode.",
        "required": [
          "certification_id",
          "certification_version",
          "mode"
        ],
        "properties": {
          "certification_id": {
            "type": "string",
            "description": "Certification identifier."
          },
          "certification_version": {
            "type": "string",
            "description": "Certification version identifier."
          },
          "device_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Device/install context. Ownership always comes from the authenticated\nuser; this value is never used as an authorization proof."
          },
          "domain_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Domain to scope a domain quiz to. Ignored for other modes."
          },
          "mode": {
            "$ref": "#/components/schemas/QuizMode",
            "description": "Quiz mode deciding how the server selects questions."
          },
          "question_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Anchor question for `recommended_practice`. The server validates it\nbelongs to the certification version; it never trusts it as the whole\npractice set."
          },
          "recommendation_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Recommendation that produced this mission, when it was recommended.\nContext only, never an authorization key."
          },
          "task_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Task to scope a task practice to. Required for `task_practice`."
          }
        }
      },
...
      "MissionResponse": {
        "type": "object",
        "description": "A server-issued mission with its questions.",
        "required": [
          "id",
          "device_id",
          "certification_id",
          "certification_version",
          "content_version",
          "mode",
          "issued_at",
          "expires_at",
          "questions"
        ],
        "properties": {
          "certification_id": {
            "type": "string",
            "description": "Certification identifier."
          },
          "certification_version": {
            "type": "string",
            "description": "Certification version identifier."
          },
          "content_version": {
            "type": "string",
            "description": "Immutable content version."
          },
          "device_id": {
            "type": "string",
            "format": "uuid",
            "description": "Device the mission belongs to."
          },
          "domain_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Domain covered, for domain quizzes."
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "description": "Expiry time."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Mission identifier."
          },
          "issued_at": {
            "type": "string",
            "format": "date-time",
            "description": "Issue time."
          },
          "mode": {
            "$ref": "#/components/schemas/QuizMode",
            "description": "Quiz mode."
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionView"
            },
            "description": "Questions in presentation order."
          },
          "task_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Task covered, for task practice."
          }
        }
      },
...
      "MissionStatus": {
        "type": "string",
        "description": "Lifecycle of a server-issued mission.",
        "enum": [
          "issued",
          "completed"
        ]
      },
...
      "PracticeTestListResponse": {
        "type": "object",
        "description": "Practice tests available for one certification.",
        "required": [
          "practice_tests"
        ],
        "properties": {
          "practice_tests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PracticeTestSummaryDto"
            },
            "description": "Available practice tests."
          }
        }
      },
      "PracticeTestResponse": {
        "type": "object",
        "description": "Learner-safe practice-test content served before submission.",
        "required": [
          "id",
          "title",
          "exam_code",
          "certification_version",
          "content_version",
          "time_limit_minutes",
          "question_count",
          "scored_question_count",
          "question_types",
          "items"
        ],
        "properties": {
          "certification_version": {
            "type": "string",
            "description": "Certification version identifier."
          },
          "content_version": {
            "type": "string",
            "description": "Immutable content version."
          },
          "exam_code": {
            "type": "string",
            "description": "Official exam code."
          },
          "id": {
            "type": "string",
            "description": "Stable practice-test identifier."
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PracticeTestItemView"
            },
            "description": "Items in authored presentation order."
          },
          "question_count": {
            "type": "integer",
            "description": "Total authored items.",
            "minimum": 0
          },
          "question_types": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Response types present."
          },
          "scored_question_count": {
            "type": "integer",
            "description": "Items that count toward the practice score.",
            "minimum": 0
          },
          "time_limit_minutes": {
            "type": "integer",
            "format": "int64",
            "description": "Exam time limit in minutes."
          },
          "title": {
            "type": "string",
            "description": "Learner-facing title."
          }
        }
      },
      "PracticeTestResultResponse": {
        "type": "object",
        "description": "Full practice-test result and review, available only after submission.",
        "required": [
          "id",
          "title",
          "exam_code",
          "total_questions",
          "scored_question_count",
          "correct_count",
          "raw_accuracy",
          "answered_count",
          "unanswered_count",
          "domain_breakdown",
          "questions",
          "score_note"
        ],
        "properties": {
          "answered_count": {
            "type": "integer",
            "description": "Items with a submitted answer.",
            "minimum": 0
          },
          "correct_count": {
            "type": "integer",
            "description": "Scored items answered correctly.",
            "minimum": 0
          },
          "domain_breakdown": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PracticeTestDomainResult"
            },
            "description": "Per-domain scored accuracy, derived from authored content."
          },
          "exam_code": {
            "type": "string",
            "description": "Official exam code."
          },
          "id": {
            "type": "string",
            "description": "Practice-test identifier."
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PracticeTestItemResult"
            },
            "description": "Per-item review in authored order."
          },
          "raw_accuracy": {
            "type": "number",
            "format": "double",
            "description": "Raw accuracy across scored items, in `[0, 1]`."
          },
          "score_note": {
            "type": "string",
            "description": "Explicit note that this raw score is not an AWS scaled score."
          },
          "scored_question_count": {
            "type": "integer",
            "description": "Items that count toward the practice score.",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "Learner-facing title."
          },
          "total_questions": {
            "type": "integer",
            "description": "Total authored items.",
            "minimum": 0
          },
          "unanswered_count": {
            "type": "integer",
            "description": "Items without a submitted answer.",
            "minimum": 0
          }
        }
      },
      "PracticeTestSubmissionRequest": {
        "type": "object",
        "description": "One-shot submission of a practice-test attempt.\n\nAnswers are keyed by question id; unanswered items are simply omitted.",
        "properties": {
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PracticeTestAnswerRequest"
            },
            "description": "Submitted answers, at most one per question."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}